Update jquery.address.d.ts
Added the complete jQuery address API to the definition file. Reordered them to be consistent with the documentation (http://www.asual.com/jquery/address/docs/) and included the same documentation in the code (for use with Visual Studio for example).

Also fixed the existing methods where return types in the definition file were incorrect. For example: method chaining was not supported.