Update URI.d.ts
uri.search(true) does return an object, but Typescript believes there is a difference between an Object and any.  Thus, if your querystring includes ?myParam=myValue, you should be able to access the value using uri.search(true).myParam.  However, the compiler threw the error "The property 'myParam' does not exist on on value of type 'Object'".  Changing the return type to any fixes this.