Update protobufjs.d.ts
This PR includes the [ProtoBuf.Long](https://github.com/dcodeIO/ProtoBuf.js/blob/master/src/ProtoBuf.js#L19) property.  I based the definition on the existing long.d.ts.  I didn't use it directly because I didn't want to create a dependency and because the long.d.ts would have to be changed to do this in the first place.

@panuhorsmalahti - does this look okay to you?