Make MenuOptions extend MenuEvents
The menu widget definition does not allow a MenuEvents object to be passed to it during initialization. For example, this gives an error from TypeScript:

$el.menu({select: (e, ui) => { }});

I modified the MenuOptions definition to extend MenuEvents and removed MenuEvents from the Menu interface definition. This makes it consistent with how the other widgets are defined and allows both options and events to be defined when creating a menu widget.