mongoose.d.ts Document interface should define _id as any
In the Document interface definition of mongoose.d.ts, the _id field
definition of _id: Types.ObjectId was wrong, as it can be any type. This
commit changes the definition to _id: any to allow interfaces that
extend Document to refine the definition of _id as appropriate to match
the schema of the model/collection.