The click event handler for the title action of Infobox entities needs to accept an optional MouseEvent input parameter (#9452)
Because the Infobox title click handler is attached to an anchor tag with a hash as its destination (<a href="#">), the hash can cause problems in framework like Angular when used in conjunction with a <base href="/'> tag.  To get around this, the MouseEvent object needs to be available inside the event handler so that we can call the "preventDefault()" method to prevent navigation to "{baseHref}/#".