Added Redis retry_strategy
Most of the ClientOpts that have to do with reconnecting have been
deprecated in favor of the `retry_strategy` property. The
DefinitelyTyped repo is outdated and is missing the retry_strategy
property. See the docs at https://github.com/NodeRedis/node_redis.

The following options of ClientOpts are deprecated in favor of
retry_strategy:
- retry_max_delay
- connect_timeout
- max_attempts