node: pause() and resume() return `this`
Readable streams and streamlike interfaces in node have pause() and resume() methods. They're defined to return `this` in the documentation: https://nodejs.org/api/stream.html#stream_readable_pause

This is relevant when someone extends a stream, as the pause and resume methods on a subtype should return the subtype, not the super type.