sip.js: fix importing as a module (#16811)
* sip.js: fix conflicts with importing the SIPjs library.

- the 'sipjs' namespace is now linked to the 'sip.js' module import
- the old 'SIP' interface has been merged into the type definition in
  the 'sipjs' namespace so typing can be used naturally.

Tested successfully with typescript 2.3.3 and sip.js 0.7.7

* sip.js: fix SIP.C namespace

It was exposed as a class but it is actually a namespace containing
constants. This was causing build errors.

Also, the namespace name was wrong.

* sip.js: remove the wrapper sipjs namespace and export things instead

Following the suggestion of @andy-ms, this commits removes the wrapper
namespace and the `declare module` block, and instead just exports all
the things individually.