[react-redux] Update types for react-redux (#18510)
The types were failing to remove props they injected from requirements
for the parents rendering them. By using Omit we can omit props that are
injected from the connect component so that we don't get errors during
parent rendering.

All of this is done through type inference so that end users don't have
to be aware that this is happening.