Add Uint8Array to MIDIOutput send
The spec for Web MIDI API MIDIOutput.send at:
https://webaudio.github.io/web-midi-api/#dom-midioutput-send
suggests that both number[] and Uint8Array are acceptable types
for sending data on a MIDI output port.

The relevant text that allows for Uint8Array states:
"... while still enabling use of Uint8Arrays for efficiency in large ..."

An obvious use case is to forward MIDI events received on MIDIInputs as
these are already in the form of a Uint8Array types.

Tested with Chrome 60.0.3112.113 on Mac OS 10.12.6.