Updated dc for my project that uses composite charts and certain features of color charts
To update a project from javascript to typescript I hit a few issues with Composite charts and color charts that I could only fix by editing the types definitions.
I added links to the documentation that I used to determine the types that are required.
dc.version was added which is documented in the source code.