[snapsvg] Make handler optional for un... methods (#27252)
For `unclick` the handler argument was already marked as optional. As in snapsvg's [implementation of the `un...` methods](https://github.com/adobe-webplatform/Snap.svg/blob/master/src/mouse.js#L340) there is no difference in how the various events are handled, the handler should also be optional for the other event types. I tried this with `undblclick` in particular, and it works as expected (removes all `dblclick` events).