[@ember/routing] Fix redirect hook
Prior to this commit `redirect()` hook was mistakenly
typed and documented the same way as `refresh()`.

It's now fixed for v2 and v3 now.

While v3 is provided with a positive and a negative test
cases, v2 is only provided with a positive one.

This is because v2 uses typescript of version 2.4 which is
not able to catch the failing test case.