Allow both React nodes and strings to be passed as TextProps
Looking at the library it's perfectly fine to pass a React node as a TextProp.
The only exception to this is rowsText, since it's used inside of an
<option> element where only a string can be passed.
Also, because noDataText is wrapped in 'normalizeComponent' method, it
can be passed both as a React component as well as a React node.