[@types/luxon] Fix and simplify Duration units.
The units that can be passed as keys in a an object for methods like
`plus` are the same unit that can be passed by string in methods like
`as`. And in both cases both singular and plural versions are accepted
as they [get normalized](https://github.com/moment/luxon/blob/5c449fef840f8023e2444cbb3497272078623a2c/src/duration.js#L233).