Preserve type information in Page.evaluate and friends
When the EvaluteFn is not a string, it may have type information.
We should try and use its type to inform the return type of
Page.evaluate, or whichever similar function was used.

Some functions (.evaluateHandle, .waitForFunction) return a JSHandle.
JSHandles are not parameterized, so we do not know the underlying type.
It may be worth changing that in a future PR.