Allow bank account id to be provided to tokens.create. (#40253)
tokens.create allows for [cloning payment methods][1] from one account
to another through stripe connect. When cloning a source, you do not
provide the full payment method information but just the string id for
the source in the appropriate key (card for credit cards and
bank_account for bank accounts).

This was done correctly for credit cards but was not done for bank
accounts. We are getting around this currently by using `as any` to
ignore the typescript errors and allow passing a string as the
bank_account.

I also fixed the comment that was copy/pasted from credit card options
and was incorrect when documenting bank account options.

[1]: https://stripe.com/docs/connect/cloning-saved-payment-methods