Update react-simple-maps geography type from object to any (#41624)
* Update react-simple-maps geography type from object to any

the object type is too restrictive, forcing people to use any in their code. The geography prop comes through as a very complex object, and using the object type prevents you from using it properly. If we want to allow any geography type (I am not sure if they vary by map, but I assume they would), we should type it as any, not as object.

That being said, I would happily contribute my type definition for geography if it is the same across all maps provided.

* update any to unknown

* unknown back to any for TypeScript 2 support