Updates Emscripten types with new MODULARIZE API (#44590)
* Updates Emscripten types with new `MODULARIZE` API

This corresponds with the changes to Emscripten in emscripten-core/emscripten#10697,
creating a type for the new Promise-based API of the factory function
generated when using `MODULARIZE`.

Removes the obsolete `then` method from EmscriptenModule and separates
the type for the factory function generated when using the `MODULARIZE`
build option from the rest of the `EmscriptenModule` interface. This
will make it easier to extend the interface.

Also removes the global `Module` variable declaration as this is not
applicable for every application. This was not ideal as it left a very
generic variable name in the global scope for any application that used
Emscripten, regardless of whether there was actually a global variable.
This now leaves it up to the user to declare their module instance as
appropriate for them.

Fixes the reference in sql.js, using the EmscriptenModule interface
directly.

* Runs prettier on the emscripten types