Improve react's no-return hack for TS 4.3 (#51081)
* Improve react's no-return hack

React, along with react-dom and react-test-renderer, have a few return
types that are intended to disallow returning anything except
`undefined`. Previously, the union type `void | undefined` worked for
this purpose, but with TS 4.3's new, more aggressive subtype reduction,
this is treated as just `void`. And `void`-returning signatures allow
*anything* to be returned.

The solution is to use a brand with void, like `void | { _a: never}`.
This allows functions that don't actually return anything to infer
`void`, which is assignable to `void`. Any functions with a return
statement will infer from the return statement, and that type will not
be assignable to `{ _a: never }`. Thanks to @tjjfvi for suggesting this
solution.

Exceptions are

1. returning expressions of type `undefined`
2. returning expressions of type `void`
3. casting to the brand type.

The first should be fine, the second should rarely happen by mistake,
and the third should only happen as an intentional workaround.

I'm opening this PR in order to maintain backward compatibility. It's
entirely possible that trying to restrict the return type of callbacks
is pointless, and that it's fine for callbacks to actually return
anything. If that's the case, it would be fine to simplify the return
type to just `void`.

* use unique symbol+type alias for brand